# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Riogas Gestión API",
    "description": "Catálogo de APIs de GOYA (app 3). Backend NestJS (prefijo /api) + route handlers del front Next.\nGenerado por `pnpm docs:api` con un parser estático del código; no editar a mano — las descripciones y ejemplos van en docs/api/anotaciones.yaml.",
    "version": "0.1.0"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ParticiparDto": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string"
          },
          "nombre": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "telefono": {
            "type": "string",
            "maxLength": 30,
            "pattern": "^\\D*(\\d\\D*){8,15}$",
            "description": "/\\d/ garantiza que haya al menos un dígito; el segundo @Matches cuenta\nlos dígitos del string completo (ignorando separadores) y exige 8-15."
          },
          "edad": {
            "type": "integer",
            "minimum": 1,
            "maximum": 120
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 120,
            "description": "@IsEmail acepta hasta ~254 chars y la columna es VarChar(120): sin el\nMaxLength, un email largo aborta la transacción y el participante queda\nen 'error_temporal' permanente sin saber por qué."
          },
          "deviceId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 40
          },
          "fingerprint": {
            "type": "string",
            "maxLength": 64
          },
          "idioma": {
            "type": "string",
            "maxLength": 10
          },
          "plataforma": {
            "type": "string",
            "maxLength": 60
          },
          "resolucion": {
            "type": "string",
            "maxLength": 20
          },
          "gpsLat": {
            "type": "number",
            "maximum": 90
          },
          "gpsLng": {
            "type": "number",
            "maximum": 180
          }
        },
        "required": [
          "codigo",
          "deviceId",
          "edad",
          "nombre",
          "telefono"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/sorteos/publico/estado": {
      "get": {
        "operationId": "SorteosPublicoController_estado",
        "tags": [
          "sorteos-publico"
        ],
        "parameters": [
          {
            "name": "codigo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-publico.controller.ts",
        "x-goya-auth": "x-api-key (SorteosApiKeyGuard)",
        "summary": "Estado público de un código de sorteo: si es válido, si ya participó y si ganó."
      }
    },
    "/api/sorteos/publico/participar": {
      "post": {
        "operationId": "SorteosPublicoController_participar",
        "tags": [
          "sorteos-publico"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-publico.controller.ts",
        "x-goya-auth": "x-api-key (SorteosApiKeyGuard)",
        "summary": "Registra la participación desde el formulario público (guarda IP y user-agent).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticiparDto"
              }
            }
          }
        }
      }
    },
    "/api/workbench/sugerencias": {
      "get": {
        "operationId": "WorkbenchController_listar",
        "tags": [
          "workbench"
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minConfianza",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/workbench/workbench.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Cola de sugerencias de match (deduplicación), filtrable por tipo, estado y confianza.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/workbench/sugerencias/{id}/aceptar": {
      "post": {
        "operationId": "WorkbenchController_aceptar",
        "tags": [
          "workbench"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/workbench/workbench.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Acepta una sugerencia y aplica la unificación.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/workbench/sugerencias/{id}/deshacer": {
      "post": {
        "operationId": "WorkbenchController_deshacer",
        "tags": [
          "workbench"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/workbench/workbench.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Deshace una sugerencia ya aplicada.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/workbench/sugerencias/{id}/rechazar": {
      "post": {
        "operationId": "WorkbenchController_rechazar",
        "tags": [
          "workbench"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/workbench/workbench.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Rechaza una sugerencia.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
...
}